french_data_list S3 class

The french_data_list exists to hold the results of reading the files lists of Kenneth's French data library.

It provides a method to print the objects of this class.

Properties of french_data_list

Objects of class french_data_list have:

A class attribute of french_data_list.

A base type of "list" with the following elements:

info - holds the information about when and were the information was retrieved.

files_list a tibble with with a list of files that can be downloaded, the tibble contains a name, file_url and details_url column.

Behavior of french_data_list

Pretty prints the object.