Get list of files available on Kenneth's French data library website

Extract the list of files along with a description and links to them and additional information from the data library page of Prof. Kenneth French website.

Usage

get_french_data_list(max_tries = 3, refresh = FALSE)

Arguments

max_tries

a number defining the maximum number of tries to perform when downloading the page.

refresh

logical. If TRUE re-downloads the page and overwrites the cached information. Otherwise use the cached data if a download has been done in the current session.

Value

An object of class french_data_list with the following elements:

info - holds the information about when and were the information was retrieved.

files_list a tibble with with a list of files that can be downloaded, the tibble contains a name, file_url and details_url column.

Examples

# \donttest{
files_list <- get_french_data_list()
files_list
#> 
#> ── Kenneth's French data library 
#> ℹ Information collected from: https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html on Fri Mar 22 17:54:23 2024
#> 
#> 
#> ── Files list 
#> # A tibble: 297 × 3
#>    name                                                file_url      details_url
#>    <chr>                                               <chr>         <chr>      
#>  1 Fama/French 3 Factors                               ftp/F-F_Rese… Data_Libra…
#>  2 Fama/French 3 Factors [Weekly]                      ftp/F-F_Rese… Data_Libra…
#>  3 Fama/French 3 Factors [Daily]                       ftp/F-F_Rese… Data_Libra…
#>  4 Fama/French 5 Factors (2x3)                         ftp/F-F_Rese… Data_Libra…
#>  5 Fama/French 5 Factors (2x3) [Daily]                 ftp/F-F_Rese… Data_Libra…
#>  6 Portfolios Formed on Size                           ftp/Portfoli… Data_Libra…
#>  7 Portfolios Formed on Size [ex.Dividends]            ftp/Portfoli… Data_Libra…
#>  8 Portfolios Formed on Size [Daily]                   ftp/Portfoli… Data_Libra…
#>  9 Portfolios Formed on Book-to-Market                 ftp/Portfoli… Data_Libra…
#> 10 Portfolios Formed on Book-to-Market [ex. Dividends] ftp/Portfoli… Data_Libra…
#> # ℹ 287 more rows
# }